pyproject.toml
[tool.poetry]
name = "lang-chain-demo"
version = "0.1.0"
description = ""
authors = ["Naoya Yasuda <43776161+yanskun@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
langchain = "^0.0.270"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
